Renamed basic commands module.
