+++ tests/kicker/test_basic_commands.py
